chore: move dropzone handler to hook
