+++ packages/react/src/components/Storage/StorageManager/DropZone/index.ts
@@ -1,1 +1,2 @@
 export { DropZone } from './DropZone';
+export { useDropZone } from './useDropZone';
