+++ packages/react/src/components/Storage/StorageManager/DropZone/types.ts
@@ -5,7 +5,15 @@ export interface DropZoneProps {
   children?: React.ReactNode;
   displayText: StorageManagerDisplayText;
   isLoading?: boolean;
-  onChange: (
-    event: React.DragEvent<HTMLDivElement> | React.ChangeEvent<HTMLInputElement>
-  ) => void;
+  onChange: (event: React.DragEvent<HTMLDivElement>) => void;
+}
+
+export interface UseDropZoneProps extends Pick<DropZoneProps, 'onChange'> {}
+export interface UseDropZoneReturn {
+  onDragStart: (event: React.DragEvent<HTMLDivElement>) => void;
+  onDragEnter: (event: React.DragEvent<HTMLDivElement>) => void;
+  onDragLeave: (event: React.DragEvent<HTMLDivElement>) => void;
+  onDragOver: (event: React.DragEvent<HTMLDivElement>) => void;
+  onDrop: (event: React.DragEvent<HTMLDivElement>) => void;
+  inDropZone: boolean;
 }
