+++ packages/react/src/components/Storage/StorageManager/DropZone/useDropZone.tsx
@@ -1,0 +1,43 @@
+import { useState } from 'react';
+import { UseDropZoneProps, UseDropZoneReturn } from './types';
+
+export function useDropZone({ onChange }: UseDropZoneProps): UseDropZoneReturn {
+  const [inDropZone, setInDropZone] = useState(false);
+  const onDragStart = (event: React.DragEvent<HTMLDivElement>) => {
+    event.dataTransfer.clearData();
+  };
+
+  const onDragEnter = (event: React.DragEvent<HTMLDivElement>) => {
+    event.preventDefault();
+    event.stopPropagation();
+  };
+
+  const onDragLeave = (event: React.DragEvent<HTMLDivElement>) => {
+    event.preventDefault();
+    event.stopPropagation();
+    setInDropZone(false);
+  };
+
+  const onDragOver = (event: React.DragEvent<HTMLDivElement>) => {
+    event.preventDefault();
+    event.stopPropagation();
+    event.dataTransfer.dropEffect = 'copy';
+    setInDropZone(true);
+  };
+
+  const onDrop = (event: React.DragEvent<HTMLDivElement>) => {
+    event.preventDefault();
+    event.stopPropagation();
+    setInDropZone(false);
+    onChange(event);
+  };
+
+  return {
+    onDragStart,
+    onDragEnter,
+    onDragLeave,
+    onDragOver,
+    onDrop,
+    inDropZone,
+  };
+}
